Enable NeoPixels
Initialize the code that controls the ws2812/NeoPixels.

For testing, I've made the "main" force some effect settings.  Of
course, it will have to be removed eventually.

It works (the LEDs light up), but it seems to me like the pattern is
going a bit fast, there might be some timing problems.

+++ nrf52/src/app/timer.h
@@ -12,7 +12,7 @@
 #define BATTERY_MANAGER_TIMER_TIMEOUT_MS 	1000 /* ms */
 
 void timer_init(void);
-uint64_t get_current_time_millis(void);
+uint32_t get_current_time_millis(void);
 void start_battery_status_timer(void);
 void stop_battery_status_timer(void);
 void start_battery_manage_timer(void);
